refactor: sql external system api
